% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_prior.R
\name{beta_prior}
\alias{beta_prior}
\title{Re-parametrize Mean and Standard Deviation parameters for Beta Distribution into Shape Parameters}
\usage{
beta_prior(m, sd)
}
\arguments{
\item{m}{Numeric. Mean of the Beta distribution}

\item{sd}{Positive numeric. Standard deviation of the Beta distribution}
}
\description{
Function takes mean and standard deviation as an argument and returns conventional alpha and beta shape parameters
}
\examples{

\dontrun{
# Generate hyperparameters for Stan model for Origins of Electoral Systems example
q0_alpha_beta <- mapply(m = c(q.a0 = 0.1, # Assumptions on mean of q0's
                              q.b0 = 0.1,
                              q.c0 = 0.05,
                              q.d0 = 0.3),
                        sd = rep(.01, times = 4),
                        FUN = beta_prior)

q1_alpha_beta <- mapply(m = c(q.a1 = 0.95, # Assumptions on mean of q1's
                              q.b1 = 0.9,
                              q.c1 = 0.475,
                              q.d1 = 0.5),
                        sd = rep(.01, times = 4),
                        FUN = beta_prior)

}

}
\keyword{beta}
\keyword{distribution}
