% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{biqq_binary_classes}
\alias{biqq_binary_classes}
\title{Generate Case Frequencies from Data}
\usage{
biqq_binary_classes(df, ordering = c(1, 5, 3, 7, 2, 6, 4, 8))
}
\arguments{
\item{df}{Data frame. Consists of all observed binary data}

\item{ordering}{Numeric vector. Vectors of ordering for all unique combinations in the output frequencies vector}
}
\description{
Function takes data frame and ordering as arguments and returns ordered named vector with frequencies of all unique rows in the data frame
}
