% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{biqq_E_Var}
\alias{biqq_E_Var}
\title{Function to identify possible case selection strategies and assess expected gains}
\usage{
biqq_E_Var(k = 1, XY_base = rep(1, 4), strategies = NULL,
  f = my_loss, name = "Results", biqq_model = biqq, ...)
}
\arguments{
\item{k}{A scalar saying how many clues are to be sought in total}

\item{XY_base}{is the data before considering K data}

\item{strategies}{optional: if only a subset of strategies are examined strategies can be indicated  of the form "0-0-1-1" meaning seek one clue for 10 cases and one for 11 cases, or more generally c("1-1-0-0", "0-0-1-1"). Use this option with caution to be sure that naming of strategies is correct and unique}
}
\value{
A list
}
\description{
Assesses the set of strategies that could be implemented by looking for clues in  k cases given existing X,Y data, recorded in a 4-vector, `XY_base`. The output lists the possible strategies, assesses the probability of different data realizations that coudl result from each strategy, the posterior variance that arises in each case, and the overall expected posterior variance from each strategy under consideration.
}
\examples{
biqq_E_Var(k = 1, XY_base = rep(1,4), chains = 2)
biqq_E_Var(k = 1, XY_base = rep(1,4), chains = 2, strategies = c("1-0-0-1"))
biqq_E_Var(k = 2, XY_base = rep(1,4), chains = 2, strategies = c("1-0-0-1"))
}
