% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biqq_get.R
\name{biqq_get}
\alias{biqq_get}
\title{biqq_get}
\usage{
biqq_get(model = fit_biqq, Q = NULL, digits = 2, ...)
}
\arguments{
\item{model}{a fitted model}

\item{Q}{quantity of interest expressed as function. Defaults to ATE. Can be any function of data. Alternatively some key words predefined for some quantities of interest, eg ATE, phi.}

\item{digits}{digits returned}
}
\description{
Asks for a quantity from a model
}
\examples{
model <- biqq()
biqq_get(model)
biqq_get(model, Q = function(d) c(lambda_b  = mean(d$abcd.2)))
}
\keyword{biqq}
