% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biqq_lower.R
\name{biqq_lower}
\alias{biqq_lower}
\title{biqq from lower level model, taking data as arguments}
\usage{
biqq_lower(fit = fit_biqq_lower, XY = c(0, 0, 0, 0), XYK = c(0, 0, 0, 0,
  0, 0, 0, 0), XYK_data = NULL, Y_u_prior = rep(0.5, 16),
  K_u_prior = rep(0.5, 4), pi_alpha = matrix(0.5, 16, 4),
  pi_beta = matrix(0.5, 16, 4), doubly_decisive = FALSE, iter = 1000,
  chains = 2, ...)
}
\arguments{
\item{fit}{a fitted biqq_lower model}

\item{XY}{XY data -- 4-vector  00 01 10 11}

\item{XYK}{XYK data -- 12-vector 000 001 010 011...}

\item{XYK_data}{XYK data provided as a dataframe}

\item{Y_u_prior}{Dirichlet hyperparameters on 16 u_Y types}

\item{K_u_prior}{Dirichlet hyperparameters on 4 u_K types}

\item{pi_alpha}{16 alpha Beta pramaters for pis}

\item{doubly_decisive}{if TRUE pobative value defaults to clue traces  for a and b types (not extreme 0/1)}

\item{pi_alpha}{16 beta Beta pramaters for pis}
}
\value{
A stan object
}
\description{
Prior information is provided for Y_u_prior (16 elements) and K_u_prior (4 elements)
}
\examples{
print(biqq_lower(XYK = c(0, 1, 1, 0, 1, 0, 0, 1)))
print(biqq_lower(XYK = c(0, 1, 1, 0, 1, 0, 0, 1), doubly_decisive = TRUE))

}
\keyword{biqq}
\keyword{dag}
