% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{biqq_strategies}
\alias{biqq_strategies}
\title{Asssess expected learning from a host of strategies.}
\usage{
biqq_strategies(model, operations, query, n = length(model$var_names),
  m_W = 1, m_K = 1, possible_Ws = possible_W(n, m_W),
  possible_Ks = possible_K(n, m_K), sims = 2000, round = 2, Rsq = FALSE,
  f = var, symbol_unknown = "?")
}
\arguments{
\item{model}{a fitted biqq model}

\item{m_w}{Number of Ws to look at}

\item{m_k}{Number of Ks to look at}
}
\description{
Asssess expected learning from a host of strategies.
}
