% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{biqq_world}
\alias{biqq_world}
\title{Function to generate a world from a structural model}
\usage{
biqq_world(model, U, do = NULL)
}
\arguments{
\item{model}{A model, made using biqq_model}

\item{U}{A context. Generally generated using model$P()}

\item{do}{an optional  list of do operations on nodes}
}
\description{
Function to generate a world from a structural model
}
\examples{
M <- biqq_model()
biqq_world(M, M$P())
}
