% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_biqq_init.R
\docType{data}
\name{data_biqq_init}
\alias{data_biqq_init}
\title{Stan Initialization Data for biqq}
\format{Data for biqq Stan program initialization}
\usage{
data_biqq_init
}
\description{
Stan Initialization Data for biqq
}
\keyword{datasets}
