% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_es_init.R
\docType{data}
\name{data_es_init}
\alias{data_es_init}
\title{Stan Initialization Data for the Origins of Electoral Systems Example}
\format{Data for Stan program initialization for the Origins of Electoral Systems example}
\usage{
data_es_init
}
\description{
Stan Initialization Data for the Origins of Electoral Systems Example
}
\keyword{datasets}
