% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_nr_init.R
\docType{data}
\name{data_nr_init}
\alias{data_nr_init}
\title{Stan Initialization Data for the Natural Resource Curse Example}
\format{Data for Stan program initialization for the Natural Resource Curse example}
\usage{
data_nr_init
}
\description{
Stan Initialization Data for the Natural Resource Curse Example
}
\keyword{datasets}
