% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_dag.R
\name{draw_dag}
\alias{draw_dag}
\title{Draw a DAG}
\usage{
draw_dag(x = 0:1, y = c(0, 0), names = c("X", "Y"), arcs = cbind(0,
  1), add_points = FALSE, solids = rep(1, length(x)), title = "",
  contraction = 0.1, add_functions = 0, add_functions_text = NULL,
  text_shift = 0.2 * add_points, padding = 0.5, length = 0.2,
  cex = 1, box = TRUE)
}
\arguments{
\item{x}{Vector of x coordiantes}

\item{y}{Vector of y coordinates}
}
\description{
Users privide x and y coordinates,  names, and arcs.
}
\examples{
draw_dag()
}
