% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_biqq.R
\name{init_biqq}
\alias{init_biqq}
\title{Initiation of Stan Program for BIQQ}
\usage{
init_biqq(model_code = NULL, stanmodel = NULL, data)
}
\arguments{
\item{model_code}{A character string either containing the model definition or the name of a character string object in the workspace. This parameter is used only if parameter file is not specified. When fit is specified, the model compiled previously is used so specifying model_code is ignored}

\item{stanmodel}{An object of class \code{stanmodel}}

\item{data}{A named list or environment providing the data for the model, or a character vector for all the names of objects used as data. See the Note section below}
}
\description{
The function takes the model code and data as an argument and returns stan_fit object to use in BIQQ
}
\examples{

\dontrun{

library(biqq)

# Create model fit for Natural Resource Curse example using init_biqq function
nr_fit <-
  init_biqq(model_code = stan_nr,
            data = data_nr_init)

# Print model fit for Natural Resource Curse example
nr_fit

# Create model fit for Origins of Electoral Systems example using init_biqq function
es_fit <-
  init_biqq(model_code = stan_es,
            data = data_es_init)

# Print model fit for Origins of Electoral Systems example
es_fit
}

}
\keyword{example}
\keyword{init}
\keyword{stan}
