% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{my_loss}
\alias{my_loss}
\title{Calculate a statistic givrn some model and a function f.}
\usage{
my_loss(model, L = function(D) {     var(D$abcd[, 2] - D$abcd[, 1]) })
}
\arguments{
\item{model}{a fitted biqq model}

\item{L}{a loss function, defined over the dataframe extracted from model L. Defaults to posterior variance of ATE}
}
\description{
Defaults to the posterior variance on lambda_b - lambda_a for simple biqq model
}
\examples{
model <- biqq()
my_loss(model, f = function() {var(abcd[,2])})
}
