% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{perm}
\alias{perm}
\title{Generate permutations}
\usage{
perm(v)
}
\arguments{
\item{v}{a vector.}
}
\value{
a matrix containing all permutations of the elements of vector \code{v}.
}
\description{
Generate all posible permutations of the elements of a vector \code{v}. Each row in the output matrix contains a different permutation.
}
\examples{
perm(v = c(2,2,2))
perm(v = 3:1)
}
