% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pKw}
\alias{pKw}
\title{Calculate the Probability that a binary clue K = 1, given W.
K enters as a vector of the form c(TRUE, FALSE, ...)}
\usage{
pKw(model, operations, query, sims = NULL, U = NULL, W, K)
}
\arguments{
\item{model}{A model, made using biqq_model}

\item{operations}{A set of operations}

\item{query}{A query defined on outcomes on observables, V}

\item{sims}{optional number of simulations for draws; defaults to 500 if U not defined}

\item{W}{Preexisting data}

\item{K}{Clues being looked for}
}
\description{
Calculate the Probability that a binary clue K = 1, given W.
K enters as a vector of the form c(TRUE, FALSE, ...)
}
