% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_biqq.R
\name{plot_biqq}
\alias{plot_biqq}
\title{BIQQ Model for Natural Resource Curse Example}
\usage{
plot_biqq(fit, title = "", x_fun = function(x) {     x$abcd[, 2] - x$abcd[,
  1] }, x_name = "ATE", y_fun = function(x) {     x$pi_c },
  y_name = "Assignment propensity for c types",
  object_colors = c("darkcyan", "darkred"), adjust_value = 0.8)
}
\arguments{
\item{fit}{Stan model fit object}

\item{title}{Character string of graph name. Defaults to \code{""}}

\item{x_fun}{Function which takes extracted posterior as an argument and returns values plotted on x axis}

\item{x_name}{Character string of x axis statistic name}

\item{y_fun}{Function which takes extracted posterior as an argument and returns values plotted on y axis}

\item{y_name}{Character string of y axis statistic name}

\item{object_colors}{Character vector of length 2. Colors to be used for plotting of posterior draws and scaled density respectively. Defaults to \code{object_colors = c("darkcyan", "darkred")}}

\item{adjust_value}{Scalar. Adjustment of the smoothing bandwidth, the bandwidth used is \code{adjust*bw}, where \code{bw} is a result of "nrd0" method}
}
\description{
Fit BIQQ Model for Natural Resource Curse Example
}
\examples{

\dontrun{
 nr_fit <- init_biqq(model_code = stan_nr, data = data_nr_init)
 nr_example <- biqq_nr(fit = nr_fit, chains = 4, cores = 1)

 plot_biqq(fit = nr_example, title = "Posteriors | Both clues weak")
}

}
\keyword{biqq}
\keyword{plot}
\keyword{stan}
