% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tracing_functions.R
\name{possible_K}
\alias{possible_K}
\title{Get matrix of possible Ks observation possibilities ( is observed or it is not observed)  on up to m out of n nodes}
\usage{
possible_K(n, m)
}
\arguments{
\item{model}{a fitted pt model}
}
\description{
Get matrix of possible Ks observation possibilities ( is observed or it is not observed)  on up to m out of n nodes
}
\examples{
possible_K(3,1)

}
