% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tracing_functions.R
\name{possible_W}
\alias{possible_W}
\title{Get matrix of possible Ws: the set of possible data for a case when data is sought for *up to* m out of n nodes}
\usage{
possible_W(n, m)
}
\arguments{
\item{m}{Number of W variations to make}

\item{model}{a fitted pt model}
}
\description{
Get matrix of possible Ws: the set of possible data for a case when data is sought for *up to* m out of n nodes
}
\examples{
possible_W(3, 2)
possible_W(3, 1)
}
