% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{possible_XYK}
\alias{possible_XYK}
\title{Returns a matrix of XY, XYK data}
\usage{
possible_XYK(XY_base = c(1, 1, 1, 1), k = 1)
}
\arguments{
\item{XY_base}{is the data before considering K data}

\item{k}{number of clues sought}
}
\description{
Given XY_base and number of clues sought.
Standard ordering should be : XY: 00, 01, 10, 11
NOTE TO SELF: Currently there are dupicate rows to be removed
}
\examples{
possible_XYK()
possible_XYK(XY_base  = c(1,1,1,0), k=2)

}
