% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{prob_k_seen}
\alias{prob_k_seen}
\title{Assessing likelihood of different clue search results given prior}
\usage{
prob_k_seen(model, k = c(1, 1, 1, 1))
}
\arguments{
\item{model}{Prior (or posterior, given XY)}

\item{k}{A 4-vector saying how many clues are to be sought in each case.}
}
\value{
Dataframe showing XYK combinations and the likelihood of seeing so many positive clues in each XY case
}
\description{
Assessing likelihood of different clue search results given prior
}
\examples{
model <- biqq()
prob_k_seen(model = model, k = 0:3)
}
