% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tracing_functions.R
\name{pt_designer}
\alias{pt_designer}
\title{Declare a process tracing design from a causal model}
\usage{
pt_designer(true_model, estimation_model = true_model, ops, query,
  strategy = TRUE, N = 1, sims = 2000)
}
\arguments{
\item{true_model}{A model: not available to applied researcher}

\item{estimation_model}{A model: as supposed by applied researcher}

\item{ops}{Operations that form part of the query}

\item{query}{The query}

\item{strategy}{Data gathering strateg -- a vector of TRUES or FALSES indicating which nodes are observed}

\item{N}{Population for which inferences might be made. Estimand averages over these.}

\item{sims}{Precision for posterior calculation}
}
\description{
Warning -- do not use redesign with design produced from this designer
}
\examples{
model <- pt_model()
ops = list(c(0, NA, NA, NA, NA), c(1, NA, NA, NA, NA))
query = function(x) (x[[1]][5] ==1) & (x[[2]][5] == 0)
design <- pt_designer(model, ops = ops, query = query, sims = 500)
draw_estimates(design)
}
