% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tracing_functions.R
\name{pt_strategies}
\alias{pt_strategies}
\title{Asssess expected learning from a host of strategies.}
\usage{
pt_strategies(model, ops, query, m_W = 1, m_K = 1,
  n = length(model$var_names), possible_Ws = possible_W(n, m_W),
  possible_Ks = possible_K(n, m_K), sims = 2000, round = 2,
  Rsq = FALSE, f = var, U = NULL, symbol_unknown = "?")
}
\arguments{
\item{model}{a fitted pt model}

\item{m_w}{Number of Ws to look at}

\item{m_k}{Number of Ks to look at}
}
\description{
Asssess expected learning from a host of strategies.
}
\examples{
model <- pt_model()
ops = list(c(0, NA, NA, NA, NA), c(1, NA, NA, NA, NA))
query = function(x) (x[[1]][5] ==1) & (x[[2]][5] == 0)
pt_strategies(model = model, ops = ops, query = query, sims = 100)
}
