% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tracing_functions.R
\name{pt_world}
\alias{pt_world}
\title{Function to generate a world (values for V) from a structural model}
\usage{
pt_world(model, U = NULL, do = NULL)
}
\arguments{
\item{model}{a model made using  \code{\link{pt_model}}.}

\item{U}{a context. A realization of the exogenous variables in the structural model. Generally generated using \code{model$P()}.}

\item{do}{an optional list of do operations on nodes. An intervention.}
}
\value{
a world.
}
\description{
Function to generate a world (values for V) from a structural model
}
\examples{
M <- pt_model()
pt_world(M)
pt_world(M, U = runif(5))
}
