% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tracing_functions.R
\name{pt_worlds}
\alias{pt_worlds}
\title{Function to generate a dataset of possible worlds from a structural model with V provided}
\usage{
pt_worlds(model, N = NULL, do = NULL, U = NULL)
}
\arguments{
\item{model}{a model made using  \code{\link{pt_model}}.}

\item{N}{Number of simulations}

\item{do}{an optional list of do operations on nodes. An intervention.}

\item{U}{a sim*n_nodes dataframe .}
}
\value{
a dataframe with worlds
}
\description{
Function to generate a dataset of possible worlds from a structural model with V provided
}
\examples{
M <- pt_model()
pt_worlds(M, N = 4)
pt_worlds(M, N = 4)
U = data.frame(matrix(runif(10), 2, 5))
pt_worlds(M, U = U)
}
