% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tracing_functions.R
\name{pt_worlds_plot}
\alias{pt_worlds_plot}
\title{A function to plot worlds that satisfy a query}
\usage{
pt_worlds_plot(which_worlds, col1 = "black", col2 = "pink")
}
\arguments{
\item{which_worlds}{A worlds evaluation object generated by `pt_which`.}

\item{col1}{Color parameter for query = TRUE}

\item{col2}{Color parameter for query = FALSE}
}
\description{
A function to plot worlds that satisfy a query
}
\examples{
qw <- pt_which(pt_model(),
            ops = list(c(0, NA, NA, NA, NA), c(1, NA, NA, NA, NA)),
            query = function(x) (x[[1]][5] == 1) & (x[[2]][5] == 0))
 pt_worlds_plot(qw)
}
