% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_es.R
\docType{data}
\name{stan_es}
\alias{stan_es}
\title{Model Code for the Origins of Electoral Systems Example}
\format{Stan model code for the Origins of Electoral Systems Example}
\usage{
stan_es
}
\description{
Model Code for the Origins of Electoral Systems Example
}
\keyword{datasets}
