% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_nr.R
\docType{data}
\name{stan_nr}
\alias{stan_nr}
\title{Model Code for the Natural Resource Curse Example}
\format{Stan model code for the Natural Resource Curse Example}
\usage{
stan_nr
}
\description{
Model Code for the Natural Resource Curse Example
}
\keyword{datasets}
