% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biqq_cases.R
\name{biqq_cases}
\alias{biqq_cases}
\title{Function to count the set in each permutation of X,Y,K}
\usage{
biqq_cases(D)
}
\arguments{
\item{D}{a dataframe or matrix with three columns, X, Y, K}
}
\description{
Data given in order XYK. Output summary consisent with this naming ordering from Stan Code:
 XY: order: 00, 01, 10, 11
 XYK: order 000, 001, 010,...
}
\examples{
biqq_cases(D = data.frame(X = c(0,1,1), Y = c(1,1,0), K = c(NA, NA, 1)))
}
\keyword{Conversion}
\keyword{biqq}
