% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{biqq_learning}
\alias{biqq_learning}
\title{Calculate what is learned from a set of observations, Ks, given existing data, W}
\usage{
biqq_learning(model, operations, query, sims = 2000, U = NULL,
  observed = NULL, W = NULL, clues_sought = NULL, Ks = NULL, f = var,
  add_means = FALSE, detail = FALSE, detailround = 5)
}
\arguments{
\item{model}{A model, made using biqq_model}

\item{operations}{A set of operations}

\item{query}{A query defined on outcomes on observables, V}

\item{sims}{optional number of simulations for draws; defaults to 500 if U not defined}

\item{observed}{Vector indicating what is known on each node. Values NA, 0, 1 etc}

\item{W}{Alternative argument for observed: Vector indicating what is known on each node. Values NA, 0, 1 etc. Ignored if `observed` is specified}

\item{clues_sought}{A TRUE/FALSE matrix with one column per node indicating which clues to be sought}

\item{Ks}{Alternative argument for clues_sought. TRUE/FALSE matrix with one column per node indicating which clues to be sought. Ignored if `clues_sought` is specified}

\item{f}{any function; defaults to f=var, posterior variance. For posterior mean set f = mean}

\item{detail}{defaults to FALSE. When TRUE output includes listing of possible patterns and probability of each}
}
\description{
Defaults to the posterior variance on lambda_b - lambda_a for simple biqq model
}
\examples{
model <- biqq_model()
operations = list(c(0, NA, NA, NA, NA), c(1, NA, NA, NA, NA))
query = function(x) (x[[1]][5] ==1) & (x[[2]][5] == 0)
biqq_learning(model,
              operations,
              query,
              Ks = rbind(
                     c(TRUE, FALSE, FALSE, FALSE, FALSE),
                     c(FALSE, TRUE, FALSE, FALSE, FALSE),
                     c(TRUE, TRUE, FALSE, FALSE, FALSE),
                     c(TRUE, TRUE, FALSE, TRUE, FALSE),
                     c(TRUE, TRUE, TRUE, TRUE, TRUE)
         )
)
}
