% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{biqq_learning}
\alias{biqq_learning}
\title{Calculate what is learned from a set of observations, Ks, given existing data, W}
\usage{
biqq_learning(model, operations, query_v, sims = 1000, U = NULL,
  W = rep(NA, length(model$var_names)), Ks = NULL, f = var,
  add_means = FALSE)
}
\arguments{
\item{model}{a fitted biqq model}

\item{Ks}{A matrix with one column per node indicating whther to be sought or not. Defaults to NULL}

\item{f}{any function; defaults to f=var, posterior variance. For posterior mean set f = mean}
}
\description{
Defaults to the posterior variance on lambda_b - lambda_a for simple biqq model
}
\examples{
model <- biqq_model()
operations = list(c(0, NA, NA, NA, NA), c(1, NA, NA, NA, NA))
query_v = function(x) (x[[1]][5] ==1) & (x[[2]][5] == 0)
biqq_learning(model,
              operations,
              query_v,
              Ks = rbind(
                     c(TRUE, FALSE, FALSE, FALSE, FALSE),
                     c(FALSE, TRUE, FALSE, FALSE, FALSE),
                     c(TRUE, TRUE, FALSE, FALSE, FALSE),
                     c(TRUE, TRUE, FALSE, TRUE, FALSE),
                     c(TRUE, TRUE, TRUE, TRUE, TRUE)
         )
)
}
