% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{biqq_world}
\alias{biqq_world}
\title{Function to generate a world from a structural model}
\usage{
biqq_world(model, U = model$P(), do = NULL, unlistit = TRUE)
}
\arguments{
\item{model}{a model made using  \code{\link{biqq_model}}.}

\item{U}{a context. A realization of the exogenous variables in the structural model. Generally generated using \code{model$P()}.}

\item{do}{an optional list of do operations on nodes. An intervention.}
}
\value{
a world.
}
\description{
Function to generate a world from a structural model
}
\examples{
M <- biqq_model()
biqq_world(M, M$P())
}
