% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_data.R
\docType{data}
\name{data_es}
\alias{data_es}
\title{Stan Data for the Origins of Electoral Systems Example}
\format{Data for Stan program for the Origins of Electoral Systems example}
\usage{
data_es
}
\description{
Stan Data for the Origins of Electoral Systems Example
}
\keyword{datasets}
