% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{losses}
\alias{losses}
\title{Calculate losses over multiple data possibilities}
\usage{
losses(XY_base = c(1, 1, 1, 1), k = 1, my_loss, ...)
}
\arguments{
\item{XY_base}{XY data prior to considering K data}

\item{k}{number of clues sought}

\item{my_loss}{A loss function}
}
\description{
Calculate losses over multiple data possibilities
}
\examples{
my_loss <- function(model){
D <-extract(model, pars='abcd')
return(var(D$abcd[,2]-D$abcd[,1]))
}
losses(my_loss = my_loss)
}
