% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{possible_K}
\alias{possible_K}
\title{Asssess expected learning from a host of strategies.}
\usage{
possible_K(n, m)
}
\arguments{
\item{model}{a fitted biqq model}
}
\description{
Asssess expected learning from a host of strategies.
}
