% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{possible_W}
\alias{possible_W}
\title{Get matrix of possible Ws,}
\usage{
possible_W(n, m)
}
\arguments{
\item{m}{Number of W variations to make}

\item{model}{a fitted biqq model}
}
\description{
Get matrix of possible Ws,
}
